# server.js
import express from "express";
import http from "http";
import { Server } from "socket.io";
import cors from "cors";
import dotenv from "dotenv";
import { createRequire } from "module";
import connectDB from "./config/db.js";
import { seedAdmins } from "./config/seed.js";

const require = createRequire(import.meta.url);
const swaggerDoc = require("./swagger.json");

import authRoutes from "./routes/auth.routes.js";
import usersRoutes from "./routes/users.routes.js";
import chatRoutes from "./routes/chat.routes.js";
import inviteRoutes from "./routes/invite.routes.js";
import initSocket from "./socket/socket.js";

dotenv.config();

const app = express();
const server = http.createServer(app);

/* ================= MIDDLEWARE ================= */

const corsOptions = {
  origin: true,
  credentials: true,
  methods: ["GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"],
  allowedHeaders: ["Content-Type", "Authorization"],
};

app.use(cors(corsOptions));
app.options("*", cors(corsOptions));
app.use(express.json());
app.use(express.urlencoded({ extended: true }));

/* ================= ROUTES ================= */

app.get("/", (_, res) => {
  res.status(200).json({
    status: "success",
    message: "Backend ishlayapti 🚀",
  });
});

app.get("/api/v1/api.json", (_, res) => res.json(swaggerDoc));

app.use("/api/v1/auth", authRoutes);
app.use("/api/v1/users", usersRoutes);
app.use("/api/v1/chats", chatRoutes);
app.use("/api/v1/invites", inviteRoutes);

/* ================= ERROR HANDLERS ================= */

app.use((req, res) => {
  res.status(404).json({
    status: "fail",
    message: "Route topilmadi ❌",
  });
});

app.use((err, req, res, next) => {
  console.error(err.stack);
  res.status(500).json({
    status: "error",
    message: "Server xatosi",
  });
});

/* ================= SERVER ================= */

const PORT = process.env.PORT || 5001;

const bootstrap = async () => {
  await connectDB();
  await seedAdmins();

  const io = new Server(server, {
    cors: {
      origin: true,
      credentials: true,
      methods: ["GET", "POST"],
      allowedHeaders: ["Content-Type", "Authorization"],
    },
  });

  initSocket(io);

  server.listen(PORT, () => {
    console.log(`🚀 Server running on http://localhost:${PORT}`);
  });
};

bootstrap().catch(err => {
  console.error("❌ Startup error:", err);
  process.exit(1);
});