# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RTK Chat App API",
    "description": "Real-time chat application API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://rtk-chat-backend.onrender.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:5001",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "username": { "type": "string" },
          "role": { "type": "string", "enum": ["user", "admin", "superAdmin"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Chat": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "title": { "type": "string" },
          "participants": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/User" }
          },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Invite": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "chatId": { "$ref": "#/components/schemas/Chat" },
          "from": { "$ref": "#/components/schemas/User" },
          "to": { "$ref": "#/components/schemas/User" },
          "status": { "type": "string", "enum": ["pending", "accepted", "declined"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Foydalanuvchilar ro'yxati",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 5 } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Foydalanuvchilar",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/User" }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": { "type": "integer" },
                        "limit": { "type": "integer" },
                        "total": { "type": "integer" },
                        "totalPages": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Yangi foydalanuvchi qo'shish",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username"],
                "properties": {
                  "username": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Yaratildi",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/chats": {
      "get": {
        "tags": ["Chats"],
        "summary": "Mening chatlarim",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Chatlar ro'yxati",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Chat" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Chats"],
        "summary": "Yangi chat yaratish",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": { "type": "string", "description": "Private chat uchun" },
                  "title": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat yaratildi",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Chat" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/invites": {
      "get": {
        "tags": ["Invites"],
        "summary": "Mening invitlarim",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Invitlar ro'yxati",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Invite" }
                }
              }
            }
          }
        }
      }
    },
...
  }
}